fix review remarks
