+++ ydb/core/audit/audit_log.h
@@ -38,8 +38,10 @@ namespace NActors {
 
 namespace NKikimr::NAudit {
 
+using TAuditLogParts = TVector<std::pair<TString, TString>>;
+
 extern std::atomic<bool> AUDIT_LOG_ENABLED;
 
-void SendAuditLog(const NActors::TActorSystem* sys, TVector<std::pair<TString, TString>>&& parts);
+void SendAuditLog(const NActors::TActorSystem* sys, TAuditLogParts&& parts);
 
 }   // namespace NKikimr::NAudit
