+++ ydb/core/audit/ya.make
@@ -1,7 +1,7 @@
 LIBRARY()
 
 SRCS(
-    audit_events.h
+    audit_log_item_builder.h
     audit_log.h
     audit_log_service.h
     audit_log_impl.cpp
